Move some bits around, update snaps
