+++ App.js
@@ -1,12 +1,11 @@
 import React, { useEffect } from 'react';
-import Config from 'react-native-config';
 import { MenuProvider } from 'react-native-popup-menu';
 import SplashScreen from 'react-native-splash-screen';
 
 import { Theme } from './app/constants/themes';
 import Entry from './app/Entry';
+import { FlagsProvider, flags } from './app/helpers/flags';
 import VersionCheckService from './app/services/VersionCheckService';
-import { FlagsProvider, parseFlags } from './flags';
 
 const App = () => {
   useEffect(() => {
@@ -14,11 +13,8 @@ const App = () => {
     VersionCheckService.start();
   }, []);
 
-  console.log('config', Config);
-  console.log('flags', parseFlags(Config));
-
   return (
-    <FlagsProvider flags={parseFlags(Config)}>
+    <FlagsProvider flags={flags}>
       <MenuProvider>
         <Theme use='default'>
           <Entry />
